Split out parsing logic for criticals
