+++ src/lib/browse/layers/CriticalIssuesLayerControl.svelte
@@ -1,7 +1,5 @@
 <script lang="ts">
-  import type { FeatureCollection } from "geojson";
   import type { GeoJSONSource, MapGeoJSONFeature } from "maplibre-gl";
-  import readXlsxFile from "read-excel-file";
   import { map } from "../../../stores";
   import { CollapsibleCard, ColorLegend, InteractiveLayer } from "../../common";
   import {
@@ -14,8 +12,8 @@
     cleanupSource,
     emptyGeojson,
     overwriteCircleLayer,
-    setPrecision,
   } from "../../maplibre";
+  import { parseCriticalIssuesExcel } from "../data";
 
   let fileInput: HTMLInputElement;
 
@@ -77,65 +75,9 @@
     }
   }
 
-  interface Row {
-    id: string;
-    inspector: string;
-    submission_time: Date;
-    scheme_reference: string;
-    current_design_stage: string;
-    critical_type: string;
-    lat_lon: string;
-    location_description: string;
-    notes: string;
-  }
-
-  async function parseExcel(): Promise<FeatureCollection> {
-    let mapping = {
-      ID: "id",
-      Inspector: "inspector",
-      "Submission Time": "submission_time",
-      "Please enter the Scheme Reference number, (e.g. ATF2_WYO_01)":
-        "scheme_reference",
-      "Please enter current Design Stage": "current_design_stage",
-      "Select Critical Issue type below:": "critical_type",
-      "Please Enter Latitude and Longitude \n(Right click on location in Google, left click on information to copy to clipboard, paste below)":
-        "lat_lon",
-      Column1: "location_description",
-      "Enter any additional information (e.g. any comments or notes about this critical issue)":
-        "notes",
-    };
-
-    let { rows } = await readXlsxFile(fileInput.files![0], {
-      map: mapping,
-      sheet: "Form Input",
-    });
-
-    let gj: FeatureCollection = {
-      type: "FeatureCollection",
-      features: [],
-    };
-    for (let row of rows as Row[]) {
-      let coordinates = setPrecision(
-        row.lat_lon.split(",").map(parseFloat).reverse()
-      );
-      let properties: { [name: string]: any } = row;
-      properties.submission_time = row.submission_time.toLocaleString();
-      gj.features.push({
-        type: "Feature",
-        properties,
-        geometry: {
-          type: "Point",
-          coordinates,
-        },
-        id: gj.features.length + 1,
-      });
-    }
-    return gj;
-  }
-
   async function onChange(e: Event) {
     try {
-      let gj = await parseExcel();
+      let gj = await parseCriticalIssuesExcel(fileInput.files![0]);
       ($map.getSource(source) as GeoJSONSource).setData(gj);
       numberIssues = gj.features.length;
       errorMessage = "";
